@@ -1,7 +1,6 @@
 
 package yuku.alkitab.base.ac;
 
-import android.app.AlertDialog;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
@@ -11,6 +10,8 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.ContextMenu;
 import android.view.ContextMenu.ContextMenuInfo;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
@@ -20,6 +21,7 @@ import android.widget.AdapterView.OnItemClickListener;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+import com.afollestad.materialdialogs.AlertDialogWrapper;
 import com.mobeta.android.dslv.DragSortController;
 import com.mobeta.android.dslv.DragSortListView;
 import yuku.afw.D;
@@ -50,7 +52,7 @@ public class MarkersActivity extends BaseActivity {
 	DragSortListView lv;
     View bGotoSync;
 	
-	BookmarkFilterAdapter adapter;
+	MarkerFilterAdapter adapter;
 
 	public static Intent createIntent() {
 		return new Intent(App.context, MarkersActivity.class);
@@ -62,7 +64,7 @@ public class MarkersActivity extends BaseActivity {
 		setContentView(R.layout.activity_markers);
 		setTitle(R.string.activity_title_markers);
 		
-		adapter = new BookmarkFilterAdapter();
+		adapter = new MarkerFilterAdapter();
 		adapter.reload();
 
 		lv = V.get(this, android.R.id.list);
@@ -70,7 +72,7 @@ public class MarkersActivity extends BaseActivity {
 		lv.setOnItemClickListener(lv_click);
 		lv.setAdapter(adapter);
 
-        BookmarkFilterController c = new BookmarkFilterController(lv, adapter);
+        MarkerFilterController c = new MarkerFilterController(lv, adapter);
         lv.setFloatViewManager(c);
         lv.setOnTouchListener(c);
 
@@ -80,6 +82,34 @@ public class MarkersActivity extends BaseActivity {
         bGotoSync.setOnClickListener(v -> startActivity(SyncSettingsActivity.createIntent()));
 
 		App.getLbm().registerReceiver(br, new IntentFilter(ACTION_RELOAD));
+	}
+
+	@Override
+	public boolean onCreateOptionsMenu(final Menu menu) {
+		getMenuInflater().inflate(R.menu.activity_markers, menu);
+		return true;
+	}
+
+	@Override
+	public boolean onPrepareOptionsMenu(final Menu menu) {
+		final MenuItem menuLabelSort = menu.findItem(R.id.menuLabelSort);
+
+		final int labelCount = adapter.getLabelCount();
+		menuLabelSort.setVisible(labelCount > 1);
+
+		return true;
+	}
+
+	@Override
+	public boolean onOptionsItemSelected(final MenuItem item) {
+		switch (item.getItemId()) {
+			case R.id.menuLabelSort:
+				S.getDb().sortLabelsAlphabetically();
+				adapter.reload();
+				return true;
+		}
+
+		return super.onOptionsItemSelected(item);
 	}
 
 	@Override
@@ -159,14 +189,14 @@ public class MarkersActivity extends BaseActivity {
 				S.getDb().deleteLabelById(label._id);
 				adapter.reload();
 			} else {
-				new AlertDialog.Builder(this)
-				.setMessage(getString(R.string.are_you_sure_you_want_to_delete_the_label_label, label.title, marker_count))
-				.setNegativeButton(R.string.cancel, null)
-				.setPositiveButton(R.string.delete, (dialog, which) -> {
-					S.getDb().deleteLabelById(label._id);
-					adapter.reload();
-				})
-				.show();
+				new AlertDialogWrapper.Builder(this)
+					.setMessage(getString(R.string.are_you_sure_you_want_to_delete_the_label_label, label.title, marker_count))
+					.setNegativeButton(R.string.cancel, null)
+					.setPositiveButton(R.string.delete, (dialog, which) -> {
+						S.getDb().deleteLabelById(label._id);
+						adapter.reload();
+					})
+					.show();
 			}
 			
 			return true;
@@ -207,12 +237,12 @@ public class MarkersActivity extends BaseActivity {
 		super.onActivityResult(requestCode, resultCode, data);
 	}
 
-	private class BookmarkFilterController extends DragSortController {
+	private class MarkerFilterController extends DragSortController {
 		int mDivPos;
 		int mDraggedPos;
 		final DragSortListView lv;
 
-		public BookmarkFilterController(DragSortListView lv, BookmarkFilterAdapter adapter) {
+		public MarkerFilterController(DragSortListView lv, MarkerFilterAdapter adapter) {
 			super(lv, R.id.drag_handle, DragSortController.ON_DOWN, 0);
 
 			this.lv = lv;
@@ -262,12 +292,12 @@ public class MarkersActivity extends BaseActivity {
 		}
 	}
 
-	private class BookmarkFilterAdapter extends BaseAdapter implements DragSortListView.DropListener {
+	private class MarkerFilterAdapter extends BaseAdapter implements DragSortListView.DropListener {
 		// 0. [icon] All bookmarks
 		// 1. [icon] Notes
 		// 2. [icon] Highlights
 		// 3. Unlabeled bookmarks
-		// 4. dst label2
+		// 4 and so on. labels
 
 		List<Label> labels;
 		
@@ -362,6 +392,11 @@ public class MarkersActivity extends BaseActivity {
 			}
 			
 			notifyDataSetChanged();
+			supportInvalidateOptionsMenu();
+		}
+
+		public int getLabelCount() {
+			return labels.size();
 		}
 	}
 }
