@@ -2,7 +2,6 @@ package yuku.alkitab.base.util;
 
 import android.app.Activity;
 import android.app.AlertDialog;
-import android.app.ProgressDialog;
 import android.content.Intent;
 import android.database.Cursor;
 import android.database.DatabaseUtils;
@@ -12,6 +11,8 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.Log;
 import android.util.Xml;
+import com.afollestad.materialdialogs.AlertDialogWrapper;
+import com.afollestad.materialdialogs.MaterialDialog;
 import gnu.trove.list.TIntList;
 import gnu.trove.list.array.TIntArrayList;
 import gnu.trove.map.hash.TIntLongHashMap;
@@ -123,19 +124,15 @@ public class BookmarkImporter {
 	}
 
 	public static void importBookmarks(final Activity activity, @NonNull final InputStream fis, final boolean finishActivityAfterwards) {
+		final MaterialDialog pd = new MaterialDialog.Builder(activity)
+			.content(R.string.mengimpor_titiktiga)
+			.cancelable(false)
+			.progress(true, 0)
+			.show();
+
 		new AsyncTask<Boolean, Integer, Object>() {
-			ProgressDialog pd;
 			int count_bookmark = 0;
 			int count_label = 0;
-
-			@Override
-			protected void onPreExecute() {
-				pd = new ProgressDialog(activity);
-				pd.setMessage(activity.getString(R.string.mengimpor_titiktiga));
-				pd.setIndeterminate(true);
-				pd.setCancelable(false);
-				pd.show();
-			}
 
 			@Override
 			protected Object doInBackground(Boolean... params) {
@@ -218,12 +215,12 @@ public class BookmarkImporter {
 
 				if (result instanceof Exception) {
 					Log.e(TAG, "Error when importing markers", (Throwable) result);
-					new AlertDialog.Builder(activity)
+					new AlertDialogWrapper.Builder(activity)
 						.setMessage(activity.getString(R.string.terjadi_kesalahan_ketika_mengimpor_pesan, ((Exception) result).getMessage()))
 						.setPositiveButton(R.string.ok, null)
 						.show();
 				} else {
-					final AlertDialog dialog = new AlertDialog.Builder(activity)
+					final AlertDialog dialog = new AlertDialogWrapper.Builder(activity)
 						.setMessage(activity.getString(R.string.impor_berhasil_angka_diproses, count_bookmark, count_label))
 						.setPositiveButton(R.string.ok, null)
 						.show();
